enable attachment of invoice to email
